requirements.txt
# Core dependencies
requests>=2.31.0
azure-storage-blob>=12.19.0
azure-identity>=1.15.0

# Optional: If using OpenAI SDK instead of raw API calls
# openai>=1.12.0
